Split command_parser.rs into pieces that can be individually imported. Makes the code less "busy".
