+++ crates/matrix_client/src/commands/mod.rs
@@ -4,3 +4,9 @@ pub mod ping_command;
 pub mod sourcecode_command;
 pub mod webhook;
 pub mod webhook_command;
+
+#[derive(Debug)]
+pub struct CommandMetadata {
+    pub user: String,
+    pub is_direct_message: bool,
+}
